# src/AppForSEII2526.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Purchase/GetPurchase": {
      "get": {
        "tags": [
          "Purchase"
        ],
        "operationId": "GetPurchase",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Purchase_Detail_DTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Purchase_Detail_DTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Purchase_Detail_DTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Purchase/CreatePurchase": {
      "post": {
        "tags": [
          "Purchase"
        ],
        "operationId": "CreatePurchase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Purchase_ForCreate_DTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Purchase_ForCreate_DTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Purchase_ForCreate_DTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Purchase_Detail_DTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Purchase_Detail_DTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Purchase_Detail_DTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Recibos/{id}": {
      "get": {
        "tags": [
          "Recibos"
        ],
        "operationId": "GetRepair",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Recibos": {
      "post": {
        "tags": [
          "Recibos"
        ],
        "operationId": "CreateRepair",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PaymentMethod": {
        "enum": [
          "CreditCard",
          "PayPal",
          "Cash"
        ],
        "type": "string"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "Purchase_Detail_DTO": {
        "required": [
          "customerNameSurname",
          "customerUserName",
          "deliveryAddress",
          "paymentMethod"
        ],
        "type": "object",
        "properties": {
          "deliveryAddress": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string",
            "format": "multiline"
          },
          "customerUserName": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "customerNameSurname": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string"
          },
          "purchaseItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Purchase_Item_DTO"
            },
            "nullable": true
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "TotalPrice": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "purchaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalQuantity": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Purchase_ForCreate_DTO": {
        "required": [
          "customerNameSurname",
          "customerUserName",
          "deliveryAddress",
          "paymentMethod"
        ],
        "type": "object",
        "properties": {
          "deliveryAddress": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string",
            "format": "multiline"
          },
          "customerUserName": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "customerNameSurname": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string"
          },
          "purchaseItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Purchase_Item_DTO"
            },
            "nullable": true
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "TotalPrice": {
            "type": "number",
            "format": "double",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Purchase_Item_DTO": {
        "type": "object",
        "properties": {
          "deviceID": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReceiptDetailDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "deliveryAddress": {
            "type": "string",
            "nullable": true
          },
          "operationDate": {
            "type": "string",
            "format": "date"
          },
          "totalPrice": {
            "type": "number",
            "format": "float"
          },
          "repairs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceiptItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReceiptForCreateDTO": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 20,
            "minLength": 2,
            "type": "string",
            "nullable": true
          },
          "surname": {
            "maxLength": 50,
            "minLength": 2,
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "deliveryAddress": {
            "maxLength": 100,
            "minLength": 5,
            "type": "string",
            "nullable": true
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "repairs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceiptItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReceiptItemDTO": {
        "required": [
          "modelToRepair",
          "repairName"
        ],
        "type": "object",
        "properties": {
          "repairName": {
            "maxLength": 500,
            "minLength": 3,
            "type": "string"
          },
          "scale": {
            "type": "string",
            "nullable": true
          },
          "modelToRepair": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "repairCost": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}